refactor(arc): upgrade `pubgrub` to v0.4.0
