+++ compiler/arc/Cargo.toml
@@ -10,6 +10,7 @@ doctest = false
 [dependencies]
 lume_assets = { path = "../../crates/lume_assets" }
 lume_errors = { path = "../../crates/lume_errors" }
+lume_hash = { path = "../../crates/lume_hash" }
 lume_options = { path = "../../rt/lume_options" }
 lume_session = { path = "../lume_session" }
 lume_span = { path = "../../crates/lume_span" }
@@ -17,7 +18,7 @@ lume_span = { path = "../../crates/lume_span" }
 error_snippet = { workspace = true }
 error_snippet_derive = { workspace = true }
 indexmap = { workspace = true }
-pubgrub = { version = "=0.3.0" }
+pubgrub = { version = "0.4" }
 semver = { version = "^1", features = ["serde"] }
 serde = { workspace = true, features = ["derive"] }
 toml = { version = "0.9.6" }
