+++ compiler/arc/src/lib.rs
@@ -1,16 +1,18 @@
-pub mod deps;
+pub mod cache;
 pub mod errors;
-pub mod fetch;
+pub(crate) mod manifest;
 pub(crate) mod parser;
+pub(crate) mod resolve;
 
 use std::path::Path;
 
 use error_snippet::Result;
 use lume_errors::DiagCtxHandle;
 use lume_session::{DependencyMap, FileLoader};
-use parser::PackageParser;
 
-pub use crate::fetch::{clean_local_cache_dir, local_cache_dir};
+pub use crate::cache::{clean_local_cache_dir, local_cache_dir};
+
+pub const DEFAULT_ARCFILE: &str = "Arcfile";
 
 /// Locates the [`lume_session::Package`] in the given root directory.
 ///
@@ -20,5 +22,6 @@ pub use crate::fetch::{clean_local_cache_dir, local_cache_dir};
 /// - the given path has no `Arcfile` stored within it
 /// - or the located `Arcfile` doesn't refer to a file.
 pub fn locate_package<L: FileLoader>(root: &Path, loader: &L, dcx: DiagCtxHandle) -> Result<DependencyMap> {
-    deps::build_dependency_tree(root, loader, dcx)
+    let resolver = resolve::Resolver::new(loader, dcx);
+    resolve::resolve(root, resolver)
 }
